+++ src/commands/setup.ts
@@ -25,26 +25,21 @@ export default function TheSetuper(params: TheSetuperConstructedParams) {
         !setupToUse
     ) throw new Error(`Given setup ${params.setup} is not valid! Choose from the list ${SETUPS.map((s) => s.name)}.`);
 
-    const contentToUse = setupToUse.seek === "tsconfig.json"
-        ? JSON.parse(setupToUse.content)
+    const contentToUse = (setupToUse.seek === "tsconfig.json" || setupToUse.seek === ".prettierrc")
+        ? parseJsonc(setupToUse.content)
         : setupToUse.seek === "fknode.yaml"
         ? parseYaml(setupToUse.content)
         : setupToUse.content;
     const path = JoinPaths(env.root, setupToUse.seek);
     const exists = CheckForPath(path);
 
+    const setupName = `${ColorString(setupToUse.name, "bold")} ${ColorString(setupToUse.seek, "italic")}`;
+
     if (
         !(Interrogate(
-            `Should we add the ${ColorString(setupToUse.name, "bold")} ${ColorString(setupToUse.seek, "italic")} file to ${
-                NameProject(
-                    project,
-                    "name-ver",
-                )
-            }?${
+            `Should we add the ${setupName} file to ${NameProject(project, "name")}?${
                 exists
-                    ? setupToUse.seek === "tsconfig.json"
-                        ? "\nNote: Your existing tsconfig.json will be merged with this template. Comments won't be preserved!"
-                        : `\nNote: Your existing ${setupToUse.seek} will be merged with this template. Duplications may happen.`
+                    ? `\nNote: Your existing ${setupToUse.seek} will be merged with this template. Comments won't be preserved and duplications might happen!`
                     : ""
             }`,
         ))
@@ -57,8 +52,9 @@ export default function TheSetuper(params: TheSetuperConstructedParams) {
 
     if (exists) {
         const fileContent = Deno.readTextFileSync(path);
-        if (setupToUse.seek === "tsconfig.json") {
+        if (setupToUse.seek === "tsconfig.json" || setupToUse.seek === ".prettierrc") {
             const parsedContent = parseJsonc(fileContent);
+            // TODO - respect user's indent size
             finalContent = JSON.stringify(deepMerge(contentToUse, parsedContent), undefined, 4);
         } else if (setupToUse.seek === "fknode.yaml") {
             const parsedContent = parseYaml(fileContent);
@@ -70,9 +66,9 @@ export default function TheSetuper(params: TheSetuperConstructedParams) {
     } else {
         finalContent = setupToUse.seek === "fknode.yaml"
             ? StringifyYaml(contentToUse)
-            : setupToUse.seek === "tsconfig.json"
+            : (setupToUse.seek === "tsconfig.json" || setupToUse.seek === ".prettierrc")
             ? JSON.stringify(contentToUse, undefined, 4)
-            : contentToUse.toString();
+            : (contentToUse as string).toString();
     }
 
     Deno.writeTextFileSync(